Move track_dataset into its own module
