+++ tests/OTAnalytics/application/test_eventlist.py
@@ -19,7 +19,7 @@ from OTAnalytics.domain.track import (
     Track,
     TrackId,
 )
-from OTAnalytics.domain.track_dataset import (
+from OTAnalytics.domain.track_dataset.track_dataset import (
     END_FRAME,
     END_OCCURRENCE,
     END_VIDEO_NAME,
