+++ tests/OTAnalytics/application/use_cases/test_create_intersection_events.py
@@ -23,7 +23,10 @@ from OTAnalytics.domain.section import (
     SectionType,
 )
 from OTAnalytics.domain.track import Detection, Track
-from OTAnalytics.domain.track_dataset import IntersectionPoint, TrackDataset
+from OTAnalytics.domain.track_dataset.track_dataset import (
+    IntersectionPoint,
+    TrackDataset,
+)
 from OTAnalytics.domain.types import EventType
 from tests.utils.builders.track_builder import TrackBuilder
