+++ tests/OTAnalytics/domain/test_track_repository.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, Mock, PropertyMock, call
 import pytest
 
 from OTAnalytics.domain.track import Track, TrackId
-from OTAnalytics.domain.track_dataset import TrackDataset
+from OTAnalytics.domain.track_dataset.track_dataset import TrackDataset
 from OTAnalytics.domain.track_repository import (
     TrackFileRepository,
     TrackListObserver,
