+++ tests/OTAnalytics/plugin_intersect_parallelization/test_multiprocessing.py
@@ -5,7 +5,7 @@ import pytest
 
 from OTAnalytics.domain.event import Event
 from OTAnalytics.domain.section import Section
-from OTAnalytics.domain.track_dataset import TrackDataset
+from OTAnalytics.domain.track_dataset.track_dataset import TrackDataset
 from OTAnalytics.plugin_intersect_parallelization.multiprocessing import (
     MultiprocessingIntersectParallelization,
 )
