+++ tests/utils/builders/track_dataset_provider.py
@@ -2,7 +2,7 @@ from typing import Literal
 from unittest.mock import Mock
 
 from OTAnalytics.domain.track import Track
-from OTAnalytics.domain.track_dataset import (
+from OTAnalytics.domain.track_dataset.track_dataset import (
     TRACK_GEOMETRY_FACTORY,
     FilterByClassTrackDataset,
     TrackDataset,
